Working LCD!!! With Mixed C and Assembly.
Some parts of the display Driver did not use the ldz macro.
This is the macro that will have to add an offset to all loaded values

The offset seems to be:

((final position of the assembly code in the final binary) / 2) + 2

Don't ask me why! But seems to work. Needs more tests.